requirements.txt

torch>=2.1.0
torchvision>=0.16.0

transformers==4.57.3
diffusers>=0.30.0
accelerate>=0.30.0
datasets>=2.18.0

deepspeed>=0.14.0

av>=12.0.0
pyarrow>=15.0.0

Pillow>=10.0.0

numpy>=1.24.0

PyYAML>=6.0
tqdm>=4.65.0
tabulate>=0.9.0
termcolor>=2.3.0

gradio>=4.0.0

torchcodec>=0.1.0
